[docs] Add `daisyUI v5` theme page
