+++ .changeset/slimy-glasses-guess.md
@@ -1,0 +1,5 @@
+---
+"docs": minor
+---
+
+Add `daisyUI v5` theme page
