+++ apps/docs/src/content/docs/_npm-with-themes.astro
@@ -5,6 +5,7 @@ import { LinkCard } from "@astrojs/starlight/components";
 import { THEMES, THEME_TTITLES, type Theme } from "@/shared";
 import { DEFAULT_LOCALE } from "@/i18n";
 import Npm from "@/components/npm.astro";
+import Picker from "@/components/picker.astro";
 
 const theme = (Astro.params.theme ?? "basic") as Theme;
 
@@ -24,7 +25,7 @@ function getUrl(isActive: boolean, url: string) {
 const locale = Astro.currentLocale ?? DEFAULT_LOCALE;
 ---
 
-<div class="theme-picker">
+<Picker>
   {
     THEMES.map((t) => {
       const isActive = t === theme;
@@ -38,7 +39,7 @@ const locale = Astro.currentLocale ?? DEFAULT_LOCALE;
       );
     })
   }
-</div>
+</Picker>
 
 <Npm pkg={`@sjsf/form @sjsf/ajv8-validator ajv@8 ${THEME_PACKAGES[theme]}`} />
 
@@ -46,20 +47,3 @@ const locale = Astro.currentLocale ?? DEFAULT_LOCALE;
   title={`${THEME_TTITLES[theme]} theme setup`}
   href={getRelativeLocaleUrl(locale, `/themes/${theme}/`)}
 />
-
-<style>
-  .theme-picker {
-    display: flex;
-    align-items: center;
-    width: 100%;
-  }
-  .theme-picker a {
-    flex-grow: 1;
-    padding: 0.5rem;
-    text-align: center;
-    border: 1px solid var(--sl-color-gray-5);
-  }
-  .theme-picker a[data-active="true"] {
-    background-color: var(--sl-color-gray-5);
-  }
-</style>
