+++ apps/docs/src/content/docs/themes/_daisyui5-theme.astro
@@ -1,9 +1,9 @@
 ---
-import styles from '@sjsf/daisyui-theme/styles.css?inline';
+import styles from '@sjsf/daisyui5-theme/styles.css?raw';
 
 import { ShadowHost } from "@/components/shadow";
 
-import Form from "./_daisyui-theme.svelte";
+import Form from "./_daisyui5-theme.svelte";
 ---
 
 <ShadowHost client:only="svelte" style={styles}>
