+++ apps/docs/src/content/docs/themes/_daisyui5-theme.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
   import { RawForm } from "@sjsf/form";
-  import { theme } from "@sjsf/daisyui-theme";
+  import { theme } from "@sjsf/daisyui5-theme";
 
   import { useAstro } from "@/astro.svelte";
   import { createCustomForm } from "@/components/custom-form";
 
-  import { boolean, schema, uiSchema } from "./_demo-schema";
-
+  import { boolean, arrayOfUniqueItems, schema, string, uiSchema } from "./_demo-schema";
+  
   const astro = useAstro();
 
   const form = createCustomForm({
@@ -16,12 +16,24 @@
       properties: {
         ...schema.properties,
         toggle: boolean,
-      }
+        filter: arrayOfUniqueItems,
+        cally: string,
+        pikaday: string,
+      },
     },
     uiSchema: {
       ...uiSchema,
       toggle: {
-        "ui:widget": "toggle"
+        "ui:widget": "toggle",
+      },
+      filter: {
+        "ui:widget": "filter",
+      },
+      cally: {
+        "ui:widget": "callyCalendar",
+      },
+      pikaday: {
+        'ui:widget': "pikadayCalendar"
       }
     },
   });
