+++ apps/docs/src/content/docs/themes/_demo-schema.ts
@@ -1,12 +1,27 @@
 import type { Schema, UiSchemaRoot } from "@sjsf/form";
 
+export const boolean: Schema = {
+  type: "boolean",
+};
+
+export const arrayOfUniqueItems: Schema = {
+  type: "array",
+  items: {
+    type: "string",
+    enum: ["one", "two", "three"],
+  },
+  uniqueItems: true,
+};
+
+export const string: Schema = {
+  type: "string",
+};
+
 export const schema: Schema = {
   type: "object",
   title: "Demo schema",
   properties: {
-    checkbox: {
-      type: "boolean",
-    },
+    checkbox: boolean,
     number: {
       type: "number",
       minimum: 5,
@@ -14,9 +29,7 @@ export const schema: Schema = {
     range: {
       type: "integer",
     },
-    text: {
-      type: "string",
-    },
+    text: string,
     textarea: {
       type: "string",
     },
@@ -28,45 +41,31 @@ export const schema: Schema = {
       type: "string",
       enum: ["one", "two", "three"],
     },
-    multiSelect: {
-      type: "array",
-      items: {
-        type: "string",
-        enum: ["one", "two", "three"],
-      },
-      uniqueItems: true,
-    },
+    multiSelect: arrayOfUniqueItems,
     radio: {
       type: "string",
       enum: ["one", "two", "three"],
     },
-    checkboxes: {
-      type: "array",
-      items: {
-        type: "string",
-        enum: ["one", "two", "three"],
-      },
-      uniqueItems: true,
-    },
+    checkboxes: arrayOfUniqueItems,
     array: {
       type: "array",
       items: {
         type: "string",
       },
-    }
+    },
   },
   additionalProperties: {
     type: "integer",
-  }
+  },
 };
 
 export const uiSchema: UiSchemaRoot = {
   range: {
     "ui:options": {
-      "input": {
+      input: {
         type: "range",
-      }
-    }
+      },
+    },
   },
   textarea: {
     "ui:widget": "textarea",
@@ -76,5 +75,5 @@ export const uiSchema: UiSchemaRoot = {
   },
   checkboxes: {
     "ui:widget": "checkboxes",
-  }
-}
+  },
+};
