+++ apps/docs/src/content/docs/themes/daisyui.mdx
@@ -7,10 +7,16 @@ sidebar:
 import { Code, Card, TabItem, Tabs, LinkButton } from '@astrojs/starlight/components';
 
 import Npm from '@/components/npm.astro';
+import Picker from '@/components/picker.astro';
 
 import DaisyuiTheme from './_daisyui-theme.astro';
 import daisyuiThemeCode from './_daisyui-theme.svelte?raw';
 import demoSchemaCode from './_demo-schema.ts?raw';
+
+<Picker>
+  <a href="." data-active={true}>v4</a>
+  <a href="../daisyui5/" data-active={false}>v5</a>
+</Picker>
 
 ## Installation
